+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+
+- Removed callable typehints to allow for constructor argument resolving
+
 ## [0.1.0] - 2017-01-14
 ### Added